Fix: Prevent DoS via JWKS refresh on bad signature
This commit addresses a Denial of Service (DoS) vulnerability in the `TokenValidator`.
Previously, any `BadSignatureError` would trigger a forced refresh of the JWKS from the IdP.
An attacker could exploit this by sending many invalidly signed tokens with a known `kid`,
causing the server to flood the IdP with requests and exhaust resources.

Changes:
- Modified `TokenValidator.validate_token` to inspect the token header before refreshing.
- If the token's `kid` is already present in the cached JWKS, the refresh is skipped.
- Updated `OIDCProvider` constructor to accept `float` for `cache_ttl` (fixing type hint).
- Added `tests/test_dos_protection.py` to verify the fix and prevent regression.

This change ensures that key rotation (unknown `kid`) still triggers a refresh, while
attacks using known keys do not.

+++ src/coreason_identity/oidc_provider.py
@@ -27,10 +27,10 @@ class OIDCProvider:
 
     Attributes:
         discovery_url (str): The OIDC discovery URL.
-        cache_ttl (int): The cache time-to-live in seconds.
+        cache_ttl (float): The cache time-to-live in seconds.
     """
 
-    def __init__(self, discovery_url: str, client: httpx.AsyncClient, cache_ttl: int = 3600) -> None:
+    def __init__(self, discovery_url: str, client: httpx.AsyncClient, cache_ttl: float = 3600) -> None:
         """
         Initialize the OIDCProvider.
